Voter tutorial update
This update is a small change to the Voter Tutorial content.
It aims to use the dedicated functions of the Cosmos-SDK as much as possible.
The resulting should benefits to the learning experience of the tutorial.

The changes taken are:

 * Add the voter module name to the permissions of the account module.
 * Use the dedicated method to send coins from an account to a module account.

By doing so, the recipient account is of module account type rather
than base account type. Therefore, this update should also add more consistency
to the module transaction logic.