docs: add feegrant module tutorial
This PR adds a feegrant module tutorial. A lot of the setup is taken from the authz module, thanks @ryanchristo !
We still need to wait until we have a proper tutorials page setup no the dev portal, so wasn't able to change any vuepress configs, but wanted to have this uploaded already.

Pinging @coldice so he can keep track of this PR once we have a tutorial page up in January.