Update Ethereum gas price unit
Reading through the documentation I noticed it said Ether was the unit of gas price for Ethereum transactions, which is used to convert Ether to gas. Using Ether twice in this way sounded a little weird and normally Ethereum gas prices are specified in gwei, so I thought this may have just been a typo.